+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 - added function to wait for a certain period of time
 - added Ini Parser
 - added Data Tests
+- added Base64 Encoding
 
 ## 0.1.11 - 05.10.2025
 - added Function to delete a directory